refactor: 검증 메서드 위치 수정에 따라 테스트 메서드 위치도 변경
